+++ jest.setup.js
@@ -3,6 +3,7 @@ const {
   printReceived,
   printExpected,
 } = require("jest-matcher-utils");
+const { toHaveNoViolations: axeMatchers } = require("jest-axe");
 const matchers = require("@testing-library/jest-dom/matchers");
 
 // Consider [aria-activedescendant="${id}"] #${id} as the focused element.
@@ -30,4 +31,4 @@ function toHaveFocus(element) {
   };
 }
 
-expect.extend({ ...matchers, toHaveFocus });
+expect.extend({ ...matchers, ...axeMatchers, toHaveFocus });