+++ packages/reakit/src/Dialog/__examples__/DialogWithCompositeWithTooltip/__tests__/index-test.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import { click, render } from "reakit-test-utils";
+import { click, render, axe } from "reakit-test-utils";
 import DialogWithCompositeWithTooltip from "..";
 
 test("open dialog with composite with tooltip", () => {
@@ -10,4 +10,11 @@ test("open dialog with composite with tooltip", () => {
   expect(label("Dialog")).toBeVisible();
   expect(text("item1")).toHaveFocus();
   expect(text("item1tooltip")).toBeVisible();
+});
+
+test("renders with no a11y violations", async () => {
+  const { baseElement } = render(<DialogWithCompositeWithTooltip />);
+  const results = await axe(baseElement);
+
+  expect(results).toHaveNoViolations();
 });