+++ packages/reakit/src/Dialog/__utils/useDisclosureRef.ts
@@ -8,27 +8,28 @@ export function useDisclosureRef(
   options: DialogOptions
 ) {
   const ref = React.useRef<HTMLElement | null>(null);
+  const animating = !!(options.animated && options.animating);
 
   React.useEffect(() => {
-    if (options.visible) return undefined;
+    if (options.visible || animating) return undefined;
     // We get the last focused element before the dialog opens, so we can move
     // focus back to it when the dialog closes.
     const onFocus = (event: FocusEvent) => {
       const target = event.target as HTMLElement;
-      ref.current = target;
-      if (options.unstable_disclosureRef) {
-        options.unstable_disclosureRef.current = target;
+      if ("focus" in target) {
+        ref.current = target;
+        if (options.unstable_disclosureRef) {
+          options.unstable_disclosureRef.current = target;
+        }
       }
     };
     const document = getDocument(dialogRef.current);
-    document.addEventListener("focus", onFocus, true);
-    return () => {
-      document.removeEventListener("focus", onFocus, true);
-    };
-  }, [options.visible, options.unstable_disclosureRef, dialogRef]);
+    document.addEventListener("focusin", onFocus);
+    return () => document.removeEventListener("focusin", onFocus);
+  }, [options.visible, animating, options.unstable_disclosureRef, dialogRef]);
 
   React.useEffect(() => {
-    if (!options.visible) return undefined;
+    if (!options.visible || animating) return undefined;
     // Safari and Firefox on MacOS don't focus on buttons on mouse down.
     // Instead, they focus on the closest focusable parent (ultimately, the
     // body element). This works around that by preventing that behavior and
@@ -43,7 +44,7 @@ export function useDisclosureRef(
     const disclosure = options.unstable_disclosureRef?.current || ref.current;
     disclosure?.addEventListener("mousedown", onMouseDown);
     return () => disclosure?.removeEventListener("mousedown", onMouseDown);
-  }, [options.visible, options.unstable_disclosureRef]);
+  }, [options.visible, animating, options.unstable_disclosureRef]);
 
   return options.unstable_disclosureRef || ref;
 }