+++ packages/reakit/src/Tooltip/__examples__/TooltipWithToolbar/__tests__/index-test.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import { render, press, hover } from "reakit-test-utils";
+import { render, press, hover, axe } from "reakit-test-utils";
 import CompositeWithTooltip from "..";
 
 test("show tooltip", () => {
@@ -19,4 +19,11 @@ test("show tooltip", () => {
   press.Escape();
   expect(text("item2")).toHaveFocus();
   expect(text("item3tooltip")).not.toBeVisible();
+});
+
+test("renders with no a11y violations", async () => {
+  const { baseElement } = render(<CompositeWithTooltip />);
+  const results = await axe(baseElement);
+
+  expect(results).toHaveNoViolations();
 });